Don't use size_t for jackbridge port_type_get_buffer_size

+++ source/jackbridge/JackBridge1.cpp
@@ -1581,14 +1581,14 @@ int jackbridge_port_type_size()
     return 32;
 }
 
-size_t jackbridge_port_type_get_buffer_size(jack_client_t* client, const char* port_type)
+uint32_t jackbridge_port_type_get_buffer_size(jack_client_t* client, const char* port_type)
 {
 #if defined(JACKBRIDGE_DUMMY)
 #elif defined(JACKBRIDGE_DIRECT)
     return jack_port_type_get_buffer_size(client, port_type);
 #else
     if (getBridgeInstance().port_type_get_buffer_size_ptr != nullptr)
-        return getBridgeInstance().port_type_get_buffer_size_ptr(client, port_type);
+        return static_cast<uint32_t>(getBridgeInstance().port_type_get_buffer_size_ptr(client, port_type));
 #endif
     return 0;
 }
